+++ contracts/src/main/java/com/github/bovvver/contracts/BookingDecisionMadeEvent.java
@@ -1,11 +1,12 @@
 package com.github.bovvver.contracts;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record BookingDecisionMadeEvent(
         UUID bookingId,
         UUID offerId,
         BookingDecisionStatus status,
-        Double salary
+        BigDecimal salary
 ) {
 }
