+++ offer-service/src/main/java/com/github/bovvver/offermanagment/offercreation/OfferCreatedResponse.java
@@ -1,5 +1,6 @@
 package com.github.bovvver.offermanagment.offercreation;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 import java.util.UUID;
@@ -11,7 +12,7 @@ public record OfferCreatedResponse(
         String status,
         LocationDTO location,
         Set<String> serviceCategories,
-        Double salary,
+        BigDecimal salary,
         LocalDateTime createdAt
 ) {
 }
